requirements.txt
aniso8601==8.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.2.1
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.7.1
mysqlclient==2.0.1
PyMySQL==0.10.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.25.10
Werkzeug==1.0.1
wincertstore==0.2
